src/main/java/org/example/Problem104.java
package org.example;

public class Problem104 {

    public static class TreeNode {
        int val;
        Problem104.TreeNode left;
        Problem104.TreeNode right;

        TreeNode() {
        }

        TreeNode(int val) {
            this.val = val;
        }

        TreeNode(int val, Problem104.TreeNode left, Problem104.TreeNode right) {
            this.val = val;
            this.left = left;
            this.right = right;
        }
    }

    public int maxDepth(TreeNode root) {
        if (root == null) {
            return 0;
        }

        int left = maxDepth(root.left) + 1;
        int right = maxDepth(root.right) + 1;

        return Math.max(left, right);
    }
}
